refactor: CarService에서 우승자 판별 로직 WinnerService로 분리
