+++ src/main/java/racingcar/controller/RacingCarController.java
@@ -1,6 +1,7 @@
 package racingcar.controller;
 
 import racingcar.service.CarService;
+import racingcar.service.WinnerService;
 import racingcar.service.MoveCountService;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
@@ -8,6 +9,7 @@ import racingcar.view.OutputView;
 public class RacingCarController {
     private final CarService service = new CarService();
     private final MoveCountService moveCountService = new MoveCountService();
+    private final WinnerService winnerService = new WinnerService();
 
     public void handleInputCarNames() {
         InputView.printCarNamesInputGuide();
@@ -29,7 +31,7 @@ public class RacingCarController {
             OutputView.printGameStatus(service.getCarList().getCars());
         }
 
-        service.determineWinner();
-        OutputView.printGameResult(service.getWinners());
+        winnerService.determineWinner();
+        OutputView.printGameResult(winnerService.getWinners());
     }
 }
