+++ src/main/java/racingcar/service/CarService.java
@@ -10,7 +10,6 @@ import java.util.List;
 public class CarService {
     private final NameValidationService nameValidationService;
     private final CarList carList = new CarList(new ArrayList<>());
-    private final List<String> winners = new ArrayList<>();
     private String[] carNames;
 
     public CarService() {
@@ -42,26 +41,5 @@ public class CarService {
 
     public CarList getCarList() {
         return carList;
-    }
-
-    public void determineWinner() {
-        int maxMoved = getMaxMoved();
-        for (Car car : carList.getCars()) {
-            if (car.getCarMoved() == maxMoved) {
-                winners.add(car.getCarName());
-            }
-        }
-    }
-
-    public int getMaxMoved() {
-        int maxMoved = 0;
-        for (Car car : carList.getCars()) {
-            maxMoved = Math.max(maxMoved, car.getCarMoved());
-        }
-        return maxMoved;
-    }
-
-    public List<String> getWinners() {
-        return winners;
     }
 }
